app/requirements.txt
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.7
fastapi==0.68.1
flake8==3.9.2
idna==3.3
mccabe==0.6.1
paho-mqtt==1.5.1
pycodestyle==2.7.0
pydantic==1.8.2
pyflakes==2.3.1
pymongo==3.12.0
requests==2.26.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
